+++ Program/rotation.h
@@ -35,23 +35,23 @@ public:
     Rotation();
     ~Rotation();
     void reset();
-    void set(double a);
-    double get();
+    void set(float a);
+    float get();
     
     //calculate new angle based on encoders
     void updateAngle(long diffL, long diffR);
     
     // Returns distance to goal in positive direction
-    double distanceLeft(double goal);
+    float distanceLeft(float goal);
 
     // Returns distance to goal in negative direction
-    double distanceRight(double goal);
+    float distanceRight(float goal);
 
     // Returns shortest distance to goal rotation
-    double distanceTo(double goal);
+    float distanceTo(float goal);
 
 private:
-	double angle;
+	float angle;
     
 };
